bst_serialization

+++ max_points.cpp
@@ -24,7 +24,7 @@ int max_point(vector<pair<int, int>>& points) {
 
 	for (auto x = slope_point_map.begin(); x != slope_point_map.end();x++) {
 		if (max_points < x->second.size()) {
-			max_points = x->second.size();
+			max_points = (int)x->second.size();
 		}
 	}
